requirements.txt
Django==2.1.2
djangorestframework==3.8.2
django-annoying==0.10.4
django-fixture-magic==0.1.3
Pillow==4.1.1
coverage==4.5.1
django-debug-toolbar==1.9.1
psycopg2-binary==2.7.4

ipdb
ipython